refactor: imported all components from single allComponents.js
