+++ src/components/allComponents.js
@@ -1,0 +1,8 @@
+import {FeaturedCategories} from "./FeaturedCategories"
+import { FiltersBar } from "./FiltersBar"
+import { Navbar } from "./Navbar";
+import { Footer } from "./Footer"
+import { ProductCard } from "./ProductCard";
+import {HorizontalCard} from "./HorizontalCard";
+
+export {FeaturedCategories,FiltersBar,Navbar,Footer,ProductCard,HorizontalCard};
