+++ src/utils/filterProducts.js
@@ -1,30 +1,30 @@
 
-export const filterProducts = (filterState, productList) => {
+export const filterProducts = (state, productList) => {
     //Filter-Products-by-Categories
-    if(filterState.categoryNames.length>0){
+    if(state.categoryNames.length>0){
         let prodList =[] ;
-           filterState.categoryNames.map(name=>
+           state.categoryNames.map(name=>
             prodList = prodList.concat(productList.filter((item) => item.categoryName === name)))
         productList = [...prodList];
     }        
     
     //Below for Price Range filter
         productList = productList.filter(
-            (item) => Number(item.price) < Number(filterState.maxPriceRange)
+            (item) => Number(item.price) < Number(state.maxPriceRange)
         );
      //Ratings filter   
         productList = productList.filter(
-            (item) => (item.rating) >= (filterState.ratings)
+            (item) => (item.rating) >= (state.ratings)
         );
     //Sort-By-Price
-        if (filterState.sortBy === "PRICE-LOW-TO-HIGH") {
+        if (state.sortBy === "PRICE-LOW-TO-HIGH") {
         productList = productList.sort((a, b) => a.price - b.price);
         }
-        if (filterState.sortBy === "PRICE-HIGH-TO-LOW") {
+        if (state.sortBy === "PRICE-HIGH-TO-LOW") {
         productList = productList.sort((a, b) => b.price - a.price);
         }
      //Hide-Out-Of-Stock   
-        if (filterState.removeOutOfStock) {
+        if (state.removeOutOfStock) {
         productList = productList.filter((item) => !item.isOutOfStock);
         }
